+++ str/str.go
@@ -1,6 +1,8 @@
 package str
 
 import (
+	"fmt"
+	"html/template"
 	"regexp"
 	"strings"
 )
@@ -8,6 +10,23 @@ import (
 // @Date：   2018/10/30 0030 15:14
 // @Author:  Joshua Conero
 // @Name:    字符串
+
+// 写入器导出为内容
+type WriterToContent struct {
+	content string
+}
+
+// 实现写入器语法
+func (wr *WriterToContent) Write(p []byte) (n int, err error) {
+	wr.content += string(p)
+	fmt.Println(wr.content, "l")
+	return 0, nil
+}
+
+// 获取值
+func (wr *WriterToContent) Content() string {
+	return wr.content
+}
 
 // 首字母大写
 func Ucfirst(str string) string {
@@ -44,7 +63,7 @@ func Lcfirst(str string) string {
 }
 
 // 安全字符串分割
-func SplitSafe(s, sep string) []string  {
+func SplitSafe(s, sep string) []string {
 	var dd []string
 	s = ClearSpace(s)
 	dd = strings.Split(s, sep)
@@ -54,9 +73,27 @@ func SplitSafe(s, sep string) []string  {
 // 清除空格
 func ClearSpace(s string) string {
 	s = strings.TrimSpace(s)
-	if strings.Index(s, " ") > -1{
+	if strings.Index(s, " ") > -1 {
 		spaceReg := regexp.MustCompile("\\s")
 		s = spaceReg.ReplaceAllString(s, "")
 	}
 	return s
+}
+
+// 根据 go template 模板编译后返回数据
+// 支持 template 模板语法
+func Render(tpl string, data interface{}) (string, error) {
+	var value string
+	temp, err := template.New("Render").Parse(tpl)
+	if err != nil {
+		fmt.Println(err.Error(), "ff")
+		return "", err
+	}
+	fmt.Println(tpl)
+	wr := &WriterToContent{}
+	err2 := temp.Execute(wr, data)
+	if err2 == nil {
+		return wr.Content(), nil
+	}
+	return value, err2
 }