+++ str/str_test.go
@@ -9,10 +9,25 @@ import (
 // @Author:  Joshua Conero
 // @Name:    名称描述
 
+func compareStr(expect, real string) bool {
+	if expect != real {
+		fmt.Println(expect + " VS " + real)
+		return false
+	}
+	return true
+}
+
+// 字符串比较
+func compareStrFunc(expect func() string, real string) bool {
+	s := expect()
+	fmt.Println(s, "<-")
+	return compareStr(s, real)
+}
+
 func TestUcfirst(t *testing.T) {
 	fn := func(s, sT string) {
-		if ClearSpace(s) != sT{
-			fmt.Println(s+" VS " + sT)
+		if ClearSpace(s) != sT {
+			fmt.Println(s + " VS " + sT)
 			t.Fail()
 		}
 	}
@@ -24,8 +39,8 @@ func TestUcfirst(t *testing.T) {
 
 func TestLcfirst(t *testing.T) {
 	fn := func(s, sT string) {
-		if ClearSpace(s) != sT{
-			fmt.Println(s+" VS " + sT)
+		if ClearSpace(s) != sT {
+			fmt.Println(s + " VS " + sT)
 			t.Fail()
 		}
 	}
@@ -38,33 +53,46 @@ func TestLcfirst(t *testing.T) {
 func TestClearSpace(t *testing.T) {
 	s := " s p a c e"
 	sT := "space"
-	if ClearSpace(s) != sT{
+	if ClearSpace(s) != sT {
 		t.Fail()
 	}
 	s = " "
 	sT = ""
-	if ClearSpace(s) != sT{
+	if ClearSpace(s) != sT {
 		t.Fail()
 	}
 
 	s = ""
 	sT = ""
-	if ClearSpace(s) != sT{
+	if ClearSpace(s) != sT {
 		t.Fail()
 	}
 }
-
 
 func TestSplitSafe(t *testing.T) {
 	fn := func(src, tar []string) {
 		oS := fmt.Sprintf("%v", src)
 		oT := fmt.Sprintf("%v", tar)
 		fmt.Println(" 输出格式：", oS)
-		if oS != oT{
+		if oS != oT {
 			t.Fail()
 		}
 	}
 	fn(SplitSafe("t est", ","), []string{"test"})
 	fn(SplitSafe("t est, test2, m   3", ","), []string{"test", "test2", "m3"})
 	fn(SplitSafe("t est,", ","), []string{"test", ""})
+}
+
+func TestRender(t *testing.T) {
+	fn := func(dd, real string, data interface{}) {
+		if !compareStrFunc(func() string {
+			c, _ := Render(dd, data)
+			return c
+		}, real) {
+			t.Fail()
+		}
+	}
+	fn("Tell me your name: {{.name}}", "Tell me your name: Joshua Conero", map[string]string{
+		"name": "Joshua Conero",
+	})
 }